Azure/DB/Cosmos/write-function/requirements.txt
azure-functions
pymongo[srv]